Adds Execute Folders
Adding a method that wraps and executes the `execute_folder` method,
when passed an array of “folders,” for each entry in the array.

This method differs from the style of `execute_folder` and `execute` by
including basic exception handling.

No existing functionality is altered by the addition of this method.